Challenge_1/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
requests
schedule
python-dotenv
pandas
pyarrow